Api_Class_Assignments/Saturday_Class_assignments/Class_09/Assignment_01/pyproject.toml
[project]
name = "class-09"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai-agents>=0.2.9",
    "python-decouple>=3.8",
]
